feat(github-sync): centralize load orchestration, add empty state (#85)
Removes the DEFAULT_STATIONS seed list entirely; every view showing the
station list now displays "Aucune station pour le moment" when it's
empty instead (business-specifications.md Sub-Issue C rule 7).

Adds HomePageContent.vue, which loads auth state, repo config, the
station list, default fuel type, and any remote GitHub sync once,
behind a page-level Suspense boundary in index.vue. This closes the
race where StationManagerTable.vue rendered a local station list
independently of StationPricesContent.vue's remote sync (rule 8,
spec-review.md): neither view exists in the DOM until the shared load
resolves.

Also bounds the GitHub proxy fetch to 10s via AbortController
(security-guidelines.md rule 7), so a hung response can no longer block
the whole page indefinitely now that everything waits on one shared
load.
