+++ src/components/EmptyStationsMessage.vue
@@ -1,0 +1,5 @@
+<template>
+  <p role="status" class="text-sm text-muted-foreground mb-4">
+    Aucune station pour le moment. Ajoutez-en une via le gestionnaire de stations.
+  </p>
+</template>
